+++ src/widgets/Co2Table/contexts/Co2Context/index.ts
@@ -1,0 +1,2 @@
+export { useCo2Data } from "./Co2Context";
+export { default as Co2ContextProvider } from "./Co2Provider";
